requirements.txt
aiohttp
aiosignal
annotated-types
anyio
attrs
black
certifi
charset-normalizer
click
dataclasses-json
distro
dnspython
email_validator
fastapi
fastapi-cli
frozenlist
h11
httpcore
httptools
httpx
idna
Jinja2
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
numpy
openai
orjson
packaging
pathspec
pinecone-client
pinecone-plugin-interface
platformdirs
pydantic
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
regex
requests
rich
shellingham
sniffio
SQLAlchemy
starlette
tenacity
tiktoken
tqdm
typer
typing-inspect
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
yarl